nimble/host: Use os_mbuf_pack_chains
if BLE_L2CAP_JOIN_RX_FRAGS is set to 1, use the new API to pack
mbuf chains instead of previous method. This new API will do more
copying (possibly) but will better utilize memory in that it
will use the entire data buffer of each mbuf in the chain.